requirements.txt
pygame~=2.0.1
Pillow~=8.3.1